% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_detect_sigma.R
\name{out_detect_sigma}
\alias{out_detect_sigma}
\title{Detekcja wartości odstających}
\usage{
out_detect_sigma(x, k = 3)
}
\arguments{
\item{x}{Wektor liczbowy}

\item{k}{Wartość liczbowa dodatnia, definiująca regułę k sigm}
}
\value{
Funkcja zwraca wektor z wartościami odstającymi
}
\description{
Funkcja wyznaczająca wartości odstające zgodnie z reguła k sigm
}
\examples{
set.seed(1234)
x <- rt(100, 1)
out_detect_sigma(x, k = 2)

}
