% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przelicz_miare.R
\name{przelicz_miare}
\alias{przelicz_miare}
\title{Przedstawianie za pomocą miar kuchennych ilości gramów}
\usage{
przelicz_miare(produkt, gramy, jednostka_docelowa = "szklanka")
}
\arguments{
\item{produkt}{Nazwa produktu, np. "maka", "cukier", "maslo", "mleko" albo "olej".}

\item{gramy}{Liczba gramow produktu.}

\item{jednostka_docelowa}{Jednostka, na ktora chcemy przeliczyc gramy.
Mozliwe wartosci: "szklanka", "lyzka", "lyzeczka".}
}
\value{
Ramka danych z wynikiem przeliczenia.
}
\description{
funkcja pomaga przedstawić za pomocą podstawowych miar kuchennych konkretną
ilość gramów w przypadku braku wagi kuchennej.
Funkcja przelicza podaną liczbę gramów na wybraną miarę kuchenną:
szklanki, łyżki albo łyżeczki. Przeliczenia są przybliżone.
}
\examples{
przelicz_miare("maka", 320, "szklanka")
przelicz_miare("cukier", 26, "lyzka")
przelicz_miare("maslo", 10, "lyzeczka")

}
