% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przelicz_na_gramy.R
\name{przelicz_na_gramy}
\alias{przelicz_na_gramy}
\title{Przeliczanie miar kuchennych na gramy}
\usage{
przelicz_na_gramy(produkt, ilosc, jednostka = "szklanka")
}
\arguments{
\item{produkt}{Nazwa produktu, np. "maka", "cukier", "maslo", "mleko" albo "olej".}

\item{ilosc}{Liczba jednostek, np. 2 szklanki albo 3 lyzki.}

\item{jednostka}{Jednostka miary. Mozliwe wartosci: "szklanka", "lyzka", "lyzeczka".}
}
\value{
Ramka danych z przeliczona wartoscia w gramach.
}
\description{
Funkcja przelicza miary kuchenne, takie jak szklanki, łyżki i łyżeczki,
na gramy dla wybranego produktu. Przeliczenia są przybliżone, ponieważ
waga jednej szklanki lub łyżki zależy od rodzaju produktu.
}
\examples{
przelicz_na_gramy("maka", 2, "szklanka")
przelicz_na_gramy("cukier", 3, "lyzka")
przelicz_na_gramy("maslo", 2, "lyzeczka")

}
