% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\docType{data}
\name{przepisy_przykladowe}
\alias{przepisy_przykladowe}
\title{Przykładowe przepisy kulinarne w jednostkach}
\format{
Ramka danych zawierająca 8 obserwacji i 5 zmiennych:
\describe{
\item{przepis}{Nazwa przepisu.}
\item{porcje}{Liczba porcji w przepisie bazowym.}
\item{skladnik}{Nazwa składnika.}
\item{ilosc}{Ilość danego składnika.}
\item{jednostka}{Jednostka miary składnika.}
}
}
\usage{
przepisy_przykladowe
}
\description{
Zbiór danych zawiera przykładowe przepisy kulinarne wraz z listą składników,
liczbą porcji, ilością składnika oraz jednostką miary.
}
\details{
Dane służą do pokazania działania funkcji skaluj_przepis() dostępnej w pakiecie
recipleScaleR.
}
\examples{
przepisy_przykladowe

}
\keyword{datasets}
