% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\docType{data}
\name{przepisy_w_miarach_kuchennych}
\alias{przepisy_w_miarach_kuchennych}
\title{Przepisy zapisane w miarach kuchennych}
\format{
Ramka danych zawierajaca informacje o przepisach:
\describe{
\item{przepis}{Nazwa przepisu.}
\item{porcje}{Liczba porcji w przepisie bazowym.}
\item{skladnik}{Nazwa skladnika.}
\item{ilosc}{Ilosc danego skladnika.}
\item{jednostka}{Jednostka miary skladnika.}
}
}
\usage{
przepisy_w_miarach_kuchennych
}
\description{
Zbior danych zawiera przykladowe przepisy kulinarne, w ktorych skladniki
zapisane sa w miarach kuchennych, takich jak szklanka, lyzka i lyzeczka.
}
\details{
Dane sluza do pokazania dzialania funkcji \code{przelicz_na_gramy()},
ktora przelicza miary kuchenne na gramy.
}
\examples{
przepisy_w_miarach_kuchennych

}
\keyword{datasets}
