% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skaluj_przepis.R
\name{skaluj_przepis}
\alias{skaluj_przepis}
\title{Przeliczanie liczby porcji}
\usage{
skaluj_przepis(dane, porcje_bazowe, porcje_docelowe)
}
\arguments{
\item{dane}{Ramka danych ze skladnikami przepisu.}

\item{porcje_bazowe}{Liczba porcji w przepisie bazowym.}

\item{porcje_docelowe}{Liczba porcji, którą chcemy otrzymac.}
}
\value{
Ramka danych z dodatkowa kolumna ilosc_po_przeliczeniu.
}
\description{
Funkcja przelicza ilosc składnikow z przepisu bazowego na podaną liczbę porcji,
np. przepis na 4 porcje, a użytkownik możę przeliczyc na
8 porcji, dzieki funkcji, która podwoi ilosc kazdego skladnika.
}
\details{
Dane wejsciowe powinny byc ramka danych zawierajaca kolumne
ilosc, w ktorej znajduja sie ilosci skladnikow.
}
\examples{
przepis <- data.frame(
  skladnik = c("maka", "cukier", "jajka"),
  ilosc = c(300, 100, 2),
  jednostka = c("g", "g", "szt.")
)

skaluj_przepis(przepis, porcje_bazowe = 4, porcje_docelowe = 8)

}
