% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\docType{data}
\name{przepisy_przykladowe}
\alias{przepisy_przykladowe}
\title{Przykladowe przepisy kulinarne}
\format{
Ramka danych zawierajaca 8 obserwacji i 5 zmiennych:
\describe{
\item{przepis}{Nazwa przepisu.}
\item{porcje}{Liczba porcji w przepisie bazowym.}
\item{skladnik}{Nazwa skladnika.}
\item{ilosc}{Ilosc danego skladnika.}
\item{jednostka}{Jednostka miary skladnika.}
}
}
\usage{
przepisy_przykladowe
}
\description{
Zbior danych zawiera przykladowe przepisy kulinarne wraz z lista skladnikow,
liczba porcji, iloscia skladnika oraz jednostka miary.
}
\details{
Dane sluza do pokazania dzialania funkcji \code{skaluj_przepis()} oraz
\code{przelicz_miare()} dostepnych w pakiecie \code{recipleScaleR}.
}
\examples{
przepisy_przykladowe

}
\keyword{datasets}
