% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skaluj_przepis.R
\name{skaluj_przepis}
\alias{skaluj_przepis}
\title{Przeliczanie liczby porcji}
\usage{
skaluj_przepis(dane, porcje_bazowe, porcje_docelowe)
}
\arguments{
\item{dane}{Ramka danych ze składnikami przepisu.}

\item{porcje_bazowe}{Liczba porcji w przepisie bazowym.}

\item{porcje_docelowe}{Liczba porcji, którą chcemy otrzymać.}
}
\value{
Ramka danych z dodatkową kolumną ilosc_po_przeliczeniu.
}
\description{
Funkcja przelicza ilość składników z przepisu bazowego na podaną liczbę porcji,
np. przepis na 4 porcje, a użytkownik możę przeliczyć na
8 porcji, dzięki funkcji, która podwoi ilość każdego składnika.
}
\details{
Dane wejściowe powinny być ramką danych zawierającą kolumnę
ilosc, w której znajdują się ilości składników.
}
\examples{
przepis <- data.frame(
  skladnik = c("maka", "cukier", "jajka"),
  ilosc = c(300, 100, 2),
  jednostka = c("g", "g", "szt.")
)

skaluj_przepis(przepis, porcje_bazowe = 4, porcje_docelowe = 8)

}
